// frontend/src/api/generated/models/RuleMiddleware.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RuleMiddleware = {
    /**
     * filter_channel, filter_user, match_regex, contains_word, cooldown
     *
     */
    type: string;
    settings: Record<string, any>;
};